+++ lib/karafka/web/pro/commanding.rb
@@ -16,7 +16,7 @@ module Karafka
     module Pro
       # Namespace for things related to consumers commanding (management)
       #
-      # This feature allows for basic of consumers. They can be stopped, moved to quiet or probed
+      # This feature allows for basic of consumers. They can be stopped, moved to quiet or traced
       # via the Web UI
       module Commanding
         class << self